Solution to Calculate Area and Perimeter of Rectangles

+++ Calculate Area and Perimeter of Rectangles@@67877d61dc107486ad5c274b/code.c
@@ -4,11 +4,11 @@ int NoOfRectangles(int r) {
     return r; 
 }
 
-int Area(int l, int w) {
+int Area(float l, float w) {
     return l * w;  
 }
 
-int Perimeter(int l, int w) {
+int Perimeter(float l,float w) {
     return 2 * (l + w);  
 }
 
@@ -19,8 +19,8 @@ int main() {
     for (int i = 1; i <= n; i++) {
         scanf("%d %d", &l, &w);
 
-        int area = Area(l, w);
-        int perimeter = Perimeter(l, w);
+        float area = Area(l, w);
+        float perimeter = Perimeter(l, w);
 
         printf("Rectangle %d: Area = %.2f, Perimeter = %.2f\n", i, area, perimeter);
     }
